chore: split lib.rs into several files to improve the code structure
