+++ src-tauri/src/window.rs
@@ -38,3 +38,86 @@ pub async fn destroy_child_window(handle: tauri::AppHandle) -> Result<(), String
  child_webview.close().map_err(|e| e.to_string())?;
  Ok(())
 }
+
+#[tauri::command]
+pub fn is_window_visible(window: tauri::Window) -> bool {
+    window.is_visible().unwrap_or(false)
+}
+
+#[tauri::command]
+pub async fn show_transaction_window(app_handle: tauri::AppHandle) {
+    if let Some(window) = app_handle.get_webview_window("tymt_d53_transaction") {
+        if let Err(e) = window.show() {
+            eprintln!("Failed to show window: {}", e);
+        }
+    } else {
+        eprintln!("Window 'tymt_d53_transaction' not found");
+    }
+
+    if let Some(window_to_hide) = app_handle.get_webview_window("tymtLauncher") {
+        if let Err(e) = window_to_hide.hide() {
+            eprintln!("Failed to hide window 'tymtLauncher': {}", e);
+        }
+    } else {
+        eprintln!("Window 'tymtLauncher' not found");
+    }
+
+    let item_ids = vec![
+        "showVisible".to_string(),
+        "fullscreen".to_string(),
+        "games".to_string(),
+        "wallet".to_string(),
+        "about".to_string(),
+        "signout".to_string(),
+        "quit".to_string(),
+        "disable_notifications".to_string(),
+    ];
+    let enabled = false;
+    set_tray_items_enabled(app_handle, item_ids, enabled).await
+}
+
+#[tauri::command]
+pub async fn hide_transaction_window(app_handle: tauri::AppHandle) {
+    if let Some(window) = app_handle.get_webview_window("tymt_d53_transaction") {
+        if let Err(e) = window.hide() {
+            eprintln!("Failed to hide window: {}", e);
+        }
+    } else {
+        eprintln!("Window 'tymt_d53_transaction' not found");
+    }
+
+    if let Some(window_to_hide) = app_handle.get_webview_window("tymtLauncher") {
+        if let Err(e) = window_to_hide.show() {
+            eprintln!("Failed to show window 'tymtLauncher': {}", e);
+        }
+    } else {
+        eprintln!("Window 'tymtLauncher' not found");
+    }
+
+    let item_ids = vec![
+        "showVisible".to_string(),
+        "fullscreen".to_string(),
+        "games".to_string(),
+        "wallet".to_string(),
+        "about".to_string(),
+        "signout".to_string(),
+        "quit".to_string(),
+        "disable_notifications".to_string(),
+    ];
+    let enabled = true;
+    set_tray_items_enabled(app_handle, item_ids, enabled).await
+}
+
+#[tauri::command]
+async fn set_tray_items_enabled(
+    app_handle: tauri::AppHandle,
+    item_ids: Vec<String>,
+    enabled: bool,
+) {
+    for item_id in item_ids {
+        let _ = app_handle
+            .tray_by_id(&item_id)
+            .unwrap()
+            .set_visible(enabled);
+    }
+}
